Move all the transports to the transports directory
