+++ lib/logging/logging.js
@@ -22,13 +22,13 @@ winston = require("winston");
 exports.logger = new winston.Logger({});
 
 const
-FileLogger = require('./winston-file'),
-StatsdLogger = require('./winston-statsd'),
-MetricsLogger = require('./winston-metrics'),
-CefLogger = require('./winston-cef'),
-HekaLogger = require('./winston-heka'),
-RabbitLogger = require('./winston-rabbit'),
-KpiLogger = require('./winston-kpi');
+FileLogger = require('./transports/file'),
+StatsdLogger = require('./transports/statsd'),
+MetricsLogger = require('./transports/metrics'),
+CefLogger = require('./transports/cef'),
+HekaLogger = require('./transports/heka'),
+RabbitLogger = require('./transports/rabbit'),
+KpiLogger = require('./transports/kpi');
 
 exports.logger.add(FileLogger, {});
 exports.logger.add(StatsdLogger, {});
