+++ lib/logging/transports/heka.js
@@ -6,8 +6,7 @@ const
 util = require('util'),
 winston = require('winston'),
 heka = require('heka'),
-logger = require('./logging').logger,
-config = require('../configuration'),
+config = require('../../configuration'),
 
 PREFIX = "browserid." + config.get('process_type');
