+++ lib/logging/transports/kpi.js
@@ -5,7 +5,7 @@
 const
 util = require('util'),
 winston = require('winston'),
-kpi_data = require('../kpi_data'),
+kpi_data = require('../../kpi_data'),
 MAX_QUEUE_SIZE = 1;
 
 var KpiLogger = function(options) {
