+++ lib/logging/transports/metrics.js
@@ -21,7 +21,7 @@ const
 winston = require('winston'),
 File = winston.transports.File,
 util = require('util'),
-configuration = require('../configuration'),
+configuration = require('../../configuration'),
 path = require('path'),
 mkdirp = require('mkdirp'),
 _ = require('underscore');
