+++ lib/logging/transports/statsd.js
@@ -5,8 +5,8 @@
 const
 util = require('util'),
 winston = require('winston'),
-logger = require('./logging').logger,
-config = require('../configuration'),
+logger = require('../logging').logger,
+config = require('../../configuration'),
 PREFIX = "browserid." + config.get('process_type') + ".";
 
 var StatsD = false;
